requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
tensorflow-cpu==2.13.0
pillow==9.5.0
python-multipart==0.0.6